fix: Fix sphinxcontrib.spelling caching issue
I was affected by this issue
https://stackoverflow.com/questions/17366625/sphinx-uses-old-module-code
on my local setup, which prevented me from being able to locally (even
inside the bao docker) check the spelling. The fix forces the spelling
check command to generate all outputs instead of trying to cache them.
This is a more portable behavior, as it allows to use the same tools be
used across projects transparently.

Signed-off-by: David Cerdeira <davidmcerdeira@gmail.com>